fix: explicit casting in decimals computation

+++ contracts/oracles/updatable/PythPriceFeed.sol
@@ -103,7 +103,7 @@ contract PythPriceFeed is IUpdatablePriceFeed, IPythPriceFeedExceptions {
 
         if (priceData.expo != -8) {
             if (priceData.expo > 0 || priceData.expo < -255) revert IncorrectPriceDecimals();
-            int256 pythDecimals = int256(10 ** uint32(-priceData.expo));
+            int256 pythDecimals = int256(uint256(10) ** uint32(-priceData.expo));
             price = price * DECIMALS / pythDecimals;
         }
